Node-ify nodejs api
In order to make the nodejs api conform to javascript conventions we'll use camelCasing
instead of snake_casing (the kebab-casing gets converted to snake_casing by the cljs
compiler). Also prefix private vars with an `-`.

Everything returns a promise, exceptions to this rule has a `Sync` suffix.

Also moved the Listeners section to the bottom of the file so we don't need the `declare`.