Fix ledger response parsing from cljs
This is a bug I introduced while moving from goog.net.XhrIo to axios so that our http
requests work both from the browser and from node.

The ledger returns a js object #js {:status 200 ,,, :data #js {,,,}}. This is
transformed into clojure data. In the case that the caller requested the :json
output-format I was only returning the clojure data, which caused the json serializer to
blow up.

I also had to change the fluree.db.connection/default-storage-read function to request
:json output when in cljs mode. I'm a little confused as to how this worked before, but
it appears to work now.