Fix websocket connection
The :or destructing only works when the `:publish` key is absent from the map. In this
case, it was present but its value was `nil`, so the `default-publish-fn` was not
substituted.

I wasn't sure how best to handle that case with the destructuring syntax, so I went with
a dumber way of defaulting.