Add closure-defines constant for separating browser/node code
This is a compile-time constant that guarantees dead code elimination. Which is good,
because I can't figure out a good way to do conditional requires in clojurescript, which
makes the file connection very difficult to build. Solution: put the one call site for
the file-conn (fluree.db.json-ld.api/connect with :method :file) behind a constant
conditional, and the whole file ns gets removed at build time, obviating the need for
conditional requires of the "fs" and "path" node modules.