stop injecting '$' into ledger names
Quoting @bplatz's slack messages regarding the previous behavior:

The high level problem attempting to be solved is to treat the ledger name as an
alias, and allow that name to be changed. Use case might be a Fluree server
making ledger foo/bar available to customers but needing to change something to
the DB that we don't allow (e.g. an incompatible) schema change. A customer
could create a new ledger, foo/bar2 but want to point foo/bar -> foo/bar2 so
apps don't need to change.

But, if you want to access the original foo/bar, you could always use foo/$bar
which skips the alias lookup.

But it never got fully implemented, and was really a not ideal strategy to deal
with the problem to being with.

New JSON-LD version makes explicit the ledger-id (which is content-addressed),
and an optional 'alias'. So it I think solves the issue in a better way.