Add `load` functionality to file conn
This drastically simplifies the implementation of the file connection. The Store
protocol isn't actually used at all, so I removed all of it. We may want to use it, but
until we do I'm taking it out. Now all reads require an address, not just a path.

In order to be able prevent collisions between the various ledgers' (and branches')
HEAD pointers, we need the ledger alias and branch name, so I had to update the commit
function.