Write file conn commits to <ledger>/<branch>/<commits>
It is nice from a discoverability standpoint to write a ledger branch's commits to a
unique directory instead of intermingling every ledger on the conn's commits.

In order to do this I added an arity to conn-proto/-commit that takes the db along with
the commit data. Then we can access the db's ledger and from there the alias and branch.