Fix datatype creation
We now do some coercion on values and `from-expanded` returns a tuple of coerced value
and datatype. This updates the callsites to use those values properly.

size-flakes was blowing up because it was not expecting a vector for the datatype.

I refactored `from-expanded` to remove some duplicate branching.

Coercion wasn't allowing actual booleans to pass through, so I added a case for booleans.