Make file conn addresses conn-relative
Instead of encoding the entire absolute path to the commit in the address, we instead
just encode the path from the conn's storage path. This makes the address shorter and
makes the on-disk files more portable.

Also added a :serializer to the file conn, and stubbed out the Store methods.