Query vars from .-o flake val also pass datatype,  perform IRI resolution in results
Variables being passed from where statement to where statement always contained only values. For .-o object values, certain downstream processes also need the datatype (e.g. if an IRI, we need to resolve the IRI's value if value is in 'select' statement). Now 'o' values passed through are always two-tuples: [value datatype].

This mean intermediate steps need to know if the value coming through is a two-tuple, so in the query parsing we retain the knowledge of which var is a two-tuple and perform appropriate deconstruction.

Also, because mixed-types are allowed, it could be that a object value in one where statement (e.g. ex:friend) might be passed to the subject of the next statement - but if the object value is not an IRI then it will break the following where statement. We now filter for only IRI values in the scenario.

Grouping is also an area that requires some additional consideration of the values now being passed in, and that is also adjusted.