remove index watch
This watch was just a placeholder for future watching functionality, but it's
not currently used. The watching functionality itself also encourages
propogating mutable state and complects the indexing and ledger update
operation. I think we should instead use explicit function calls with persistent
values instead of mutable state to implement this functionality when we
eventually need it.