Add ?retraction and query-db-before smartfunctions
`query-db-before` allows users to query the db before the transaction has been applied
to the db, which can be useful for enforcing preconditions to a transaction.

`?retraction` allows telling if a transaction includes any retracted flakes, useful if
users want to use a smartfunction to constrain only deletes (or vice versa).