Add cljs support for EcdsaSecp256k1RecoverySignature2020
The jld-processor/canonize function has differing calling conventions for cljs and clj - in cljs
it returns a promise, in clj it is synchronous. This wraps the calls in an core async go
block and reads the promise with the handy <p! macro.

As a note, our signature generation produces different signatures on cljs and clj, but
they are mutually readable. The tests verify that. See this issue for the details:

https://github.com/fluree/fluree.crypto/issues/20

Fixes #245