create indexer component
This is a proof-of-concept that we can get indexing and querying working with our
existing db implementation. Really the whole fluree.indexer.db namespace is kind of a
kludge to not break anything going forward, just to put a proper facade over it.

This implementation only works with a memory store at the moment, since a db object
cannot be serialized to disk. Hopefully once I expand my understanding of how loading
from disk works we'll be able to do it properly.