refactor indexer to be ledger-agnostic
Removed the ledger name from the indexer api. It was only used during init to prefix the
db-address, and I'd like to eventually move to a content-addressed index so if two dbs
have the same information, they have the same address.

Also renamed some of the core functions to better reflect what they do.