change conn api to only work with ledger-addresses
A db-address references a db at a point in time, but that is also available via a query
with a `t` key. The user shouldn't have to understand the distinction between ledgers
and dbs, and especially not between the two time models of `t` and db-address, which is
just an implementation detail of the indexer. There was no benefit for exposing this,
and the implementation of transact and query should really be symmetric.

This makes the conn api consistent in that everything deals with a ledger.

Co-authored-by: Marcela Poffald mpoffald@flur.ee