add `list` to publisher api
Now the publisher can report all of the ledgers it knows about, returning each ledger
map.

We changed the create-ledger-address function to prefix the ledger heads with head, so
we can "efficiently" list them with the store/list function, which takes a key
prefix. So now they all have a common prefix!

Co-authored-by: Marcela Poffald mpoffald@flur.ee