added `resolve` function to transactor api
We can now retrieve the commit data from the transactor.

We spent some time iterating on the commit structure, eventually deciding to favor easy
verification for commit readers by including the hash in the outer wrapper along with
the address and the data. The address encodes a lot of different information that can be
decoded by using the `fluree.common.identity/address-parts` function for convenience.

Co-authored-by: Marcela Poffald mpoffald@flur.ee