add persisted indexes to indexer
Note that this contains a change to existing code - using the new Store protocol inside
of fluree.db.storage.core. It's the same interface, but it will break existing
code. Next step is to consolidate these so no breakage occurs.

This doesn't currently survive a restart of the indexer, as the db-address ->
FlureeJsonLdDb mapping is lost.

Also doesn't support the FileStore yet - need to add the :serializer.

Co-authored-by: Marcela Poffald mpoffald@flur.ee