add naive `load` functionality to conn
Restructured the Commit to store the original tx so we can just re-stage it instead of
having to turn the asserts and retracts into flakes and merge them into the db, which we
don't currently have an api for in the indexer.

Once index loading works we will only have to re-stage any commits that are ahead of the
index.