fix `store/list` implementations
`list` only returns the keys that match the prefix, not the values associated with those
keys.

Type hinting fixed a bug on file writing. Needed to pass base-path and path separately
so the returned information is correct.

Write now returns the hash, address, and path because the index-writer needs the address
for something. Maybe we can move it out later, or make it optional.