add ledger-name arg to idxr/init
Now we can store the indexer artifacts (index files, tx-summaries) in the ledger
directory in the store.

Also got rid of tt-id for db-addresses. Now db-addresses are deterministic, with the
path following the form `<ledger-name>/<root-address>/<tx-summary-hash>`. Initially it
will look like: `ledger/init/init`. As transactions are staged, it will take the form
`ledger/init/<tx-summary-hash>`. Once the :reindex-min-bytes is hit, a new root will be
written and the address will take the form `ledger/<root-address>/<tx-summary-hash>`.

This lays the groundwork for writing tx-summaries to disk to checkpoint valid
transactions that are still in memory, so we can recover from an abrupt shutdown by
merging the assertions and retractions back in on top of the db root.