fix index resolver lookup
In fluree.db.indexer.default/refresh the index root gets it's id from the :address key
returned by store/write. I had assumed it was an address address, but it is actually
just the path. By changing the :address key to have just the store k, index lookups work
properly.

This would be a good target for future refactors - eliminating that concept of :address.