safely persist tx-summaries for rebuilding novelty
Had to go on a bit of hunt to figure out how index keys are written to the store. The
old way of doing it was to pass underscore-delimited index node keys to the store, which
would convert them into file paths. However, I don't want to use special store logic
just for index keys. Fortunately, all index node keys are prefixed with the ledger
"network", so by defining the network as the <ledgername>/index we can store all the
index data together.

Unfortunately, different types of keys figure out what the "network" is in different
ways. Garbage and root node keys take the network directly from the Db. Branch and leaf
node keys get the network from the fluree.db.index/empty-branch map, which in turn gets
the network from the :method of the Ledger used to create the Db.

Also converted the stage-db function to utilize the :content-address? option for the
store/write, which I think simplifies.

Also, trying to re-init the same db shouldn't work, so I added an error there.