load db from persisted data
There is room for more efficiencies in how we load the tx-summaries. There's also a
failing test - we don't end up with the same db-summary after loading as we had before -
the size and number of flakes are different. I'm not sure why we can't rebuild the same
flakes, but I put in a failing test for it so it can get fixed.