commit transaction before staging
Staging and indexing a transaction can take a while, certainly longer than writing the
transaction to disk so we don't lose it. Now we write to disk straightaway and update
the ledger head to reflect that we've committed the transaction before we process it.

If everything shuts down after we've pushed the commit but before we've staged the db,
we can still recover that data when we load the ledger after restart.