use simpler history range syntax
The old syntax had special a lot semantics packed into the form of the `t` value:
t: <t> -> "at t"
t: [<t>] -> "from t"
t: [<a> <b>] "from a to b"

It's not clear just by looking what the difference between, say, `[3]` and `3` are, and
so documentation would have to be consulted to interpret a query found in code.

By changing the syntax to a more expressive {from <t>, to <t>} map, it's clear to a
reader what the query intends to achieve. You need either a from or a to, and if you
leave it out from defaults to the earliest t and to defaults to the latest t.