refactor transactor to track last commit for each ledger
The transactor has just been reporting the last commit summary to whoever called it, but
if we publish a commit before it has been indexed (and implicitly validated), then we
may be holding on to garbage commits.

Now the transactor can persist transactions so we don't lose the data, but the
transactor head won't be published until it's valid (that part isn't implemented here,
just the transactor head state).

Also updated the persisted data structures to use string iris, continuing the work from
the indexer conversion.