load a ledger by checking the head commit from the txr
Instead of looking to the ledger to find the head commit, we need to look at the
transactor itself - the commit can't be published until after it's been
validated (indexed). But if we shut down before indexing is complete we still don't want
to lose the commit. Now we use the transactor head pointer for the ledger to load the
latest commit.