correct the commit schema flakes
It turns out if you have a flake with a predicate of `const/$iri` (0), the datatype of the
flake must be a `const/$xsd:string` (1). Otherwise the query select mechanism will try
to resolve the flake object's @id, but that requires the flake object to be a subject
id.

These commit schema flakes had predicate: iri and datatype iri, which was causing
queries that referenced them to blow up.