create necessary commit flakes on commit
We were not creating an @id for the commit, which was inserting nils for the commit
subject.

Also refactored the creation of commit flakes to make each group of flakes more clear
and only create new sids when we actually use them.

Updated the s+p+o query test to not expect nil subjects and control for the source of
non-determinism in a commit.

Co-authored-by: Marcela Poffald mpoffald@flur.ee