add tx-id to tx-head and use that to checkpoint index progress
This commit propagates the tx-head rename to the rest of the codebase and implements
loading with the new tx-id as the indexing checkpoint.

The tx-id is now included in the db-blocks, and if the tx-head tx-id does not match the
db-head tx-id then tx-summaries are resolved until the match is found and then
re-staged. Now we won't publish potentially invalid transactions at the ledger head.