add subscribe and unsubscribe methods to conn protocol
I initially built it in the indexer, since it's got the data that needs to be broadcast,
but I moved it to the connection because you need to subscribe to a ledger, and the
indexer doesn't know about ledgers. I thought about putting it on the publisher but the
publisher would then need to rely on an indexer.

It turns out putting it on the connection worked perfectly. The subscription state is
not persisted and is created fresh each time, so the connection itself doesn't need a
store.