revert "create standalone transactor-conn"
This reverts commit 89fb57c4cba613e712d7761ce8d824958cec7f79.

I started working on the indexer conn now and I started to doubt the wisdom of wrapping
the components in a connection. A connection is for composing multiple components, if
there's just a single component then it is not necessary - it's just pure overhead.

Single-component servers should have apis that directly call the component they are
wrapping.