use ledger-name instead of ledger-address within a conn
The <ns>:<type>:<store>:<path> style addresses are only useful for resolving different
types of components (type) and different storage methods (store), which has to happen
external to a connection. So we don't need to lug that around internally. Also,
ledger-name is a much nicer api than a long weird string.

Also renamed the publisher protocol methods - I'm not sure git is a great metaphor for
what is actually happening internally, so I think these new `publish` and `resolve`
methods don't carry so much metaphorical baggage and are a bit simpler.