fix tests
The `vocab/update-with` function didn't do what I expected - it takes a db and returns a
schema, and I was expecting it to return a db with the updated schema. Instead I swapped
it with the `vocab/update-with*` which takes a schema and some vocab flakes and returns
a new schema cache.

Also added the `test-with-files` dep to the :dev deps so I can run the test from
api-tests from the repl.