history query perf improvements
Using `group-by` instead of `sort-by` + `partition-by` can produce the assert and
retract flakes in one traversal instead of two. I added an extra subject in the test
suite to make sure that the final results still have a stable sort, and it appears that
they do - we don't care about the specific ordering, just that it's stable.

Also used <! instead of <? because the assert/retract chans won't ever get an error on
them, so the extra error handling isn't necessary.