wip properly filter out flakes from before from-t
History queries don't have a concept of "stale flakes" - they want all flakes between
`from-t` and `to-t`. The CachedTRangeResolver filters out "stale flakes" but _not_ all of
the flakes from before `from-t`.

This is a brute force remaking of the call stack so that we can get the correct behavior
from `query-range/time-range`.

Co-authored-by: Marcela Poffald <mpoffald@flur.ee>