inline query pipeline into time-range
The history query pipeline requires different `t-range` behavior within the `Resolver`
which is quite far down the call stack:

time-range -> index-range* -> resolve-flake-slices -> CachedTRangeResolver -> t-range

One option we considered was passing a `:resolver` opt down from `time-range` with a
value of `:history`, and then choosing the resolver behavior there. We decided it was
simpler to just treat the history query pipeline as a separate pipeline.

We also considered factoring the CachedTRangeResolver to implement a different
protocol besides `Resolver`, as it seems to be doing a slightly different job (filtering
node flakes) than the other Resolvers (fetching nodes). But, we couldn't think of a good
thing to use for polymorphic dispatch.

In the end, we inlined the `index-range*` and `resolve-flake-slices` functions into
`time-range` so we could provide our own resolver without having to solve the dispatch
problem. The consequence is if we fix the regular query pipeline we'll have to
duplicate that fix to the history query pipeline in the future, but maybe they won't
have the same types of bugs.

Co-authored-by: Marcela Poffald <mpoffald@flur.ee>