stop passing around async-cache closure
Since we were passing around a closure and the argument that it closes over we can stop
passing the closure and just call the function directly with the argument.

Also changed the name from "async-cache", which is named for how we get the value, to
"lru-cache", since that's the semantic behind the cache.