ensure all commit keys are present in commit query
We had to sneak the previous commit into the commit-flake-addition process in order to
get the previous commit and commit data ids.

We also had to update how we were constructing the results to not filter the iri flakes
out of the commit flakes because we need ids in the final map.

Co-authored-by: Marcela Poffald <mpoffald@flur.ee>