refactor commit details async pipeline
We changed the default sort order of the history query to be in ascending order by t -
earliest change first. This matches the default tspo sort order, so we no longer have to
sort the commit flakes. History still needs to be sorted because sometimes it uses
non-tspo indexes to satisfy the query.

Moved the async into calls as high up the call stack as we could, and removed
unnecessary go blocks.

Co-authored-by: Marcela Poffald <mpoffald@flur.ee>