add support for `sh:equals`
Because property pair constraints require looking at flakes with different predicates, we can't do the validation by looking at just one predicate-group of flakes at a time. Also, when we build the shape-map, we need a key for both predicates in the pair, so we  can find the appropriate shape regardless of whether we see the lhs or rhs predicate flake first.

Therefore, for property pair constraints we do two things:
- create a key/value pair in the shape `:property` for each "direction" of the pair
- flag those shapes as needing a second lookup (to find the other half of the pair) with a `:pair` key
- have a validation codepath that takes two sets of flakes at a time (both sides of the pair) for validation

When we encounter a shape that is a `:pair` type shape, we know to send validation down the `validate-property-pairs` path, rather than the one-predicate-at-a-time `validate-property` path.