prevent object unpacking with @set in context
In json-ld, everything is multicardinality under the covers. As a convenience, we take
predicates that have only one object and unwrap them from their vectors. If a user wants
to retain the vector that the object sits in regardless of the number of objects, they
can use the context to specify {@container: @set} (or @list) for that predicate.

fixes #390