only utilize conn default context on ledger create
We were merging the ledger context into the conn context when loading a ledger, since
`load` calls into `create`. This commit changes the `create` function to only pass along
the context it was provided with unchanged.

We really need to stop merging contexts, as that's not how contexts work, and we need to
stop normalizing them to keywords, but that's a change for another day.