Do not treat pattern as type `:class` if object is a var
The `:class` pattern type is for patterns that have an actual value for the class. This value is used to calculate the subclasses of that class and retrieve the relevant flakes for those subclasses. If we have a var instead of a value, we should not process the pattern as a `:class` pattern.