throw an error if attempting to wrap a db in policy twice
If a user tries to supply an identity via `:opts`, and perform an operation on a db that already has policy applied (via `wrap-policy`), this is an error.

In the future, we could maybe do something mores sophisticated, such as allowing this in cases where the "latter" policy is more restrictive, but for now this is an error.