handle string variables in values clause
If you use a string variable binding in the values clause, it was never parsed as a
variable when being used as a key in the solution map. However, it was parsed as a
variable when creating where patterns. The consequence of this is that when we called
`fluree.db.query.exec.where/assign-matched-values`, the existing solution didn't match
up with the where clause and the values weren't being used.

This commit makes sure to parse the variables in a values clause so that we are always
dealing with symbol variables internally.

Fixes #474