Do not retract internal iri flakes
We have flakes which map subject ids to iris, and they are used internally to look up iris for subjects. While it is safe to retract them in some cases (eg if no other nodes refer to that subject), they are an internal optimization and they don't *need* to be deleted.

In order to delete them only when it's safe, we would probably need to add an index lookup here. And to ensure they don't get reintroduced upon load, we would need to update our loading pipeline to handle retracting them as well. We can do that as future work, as it would save us from having these flakes hanging around when they're not needed.