update api to allow credential-wrapped requests
- stage
- query
- multiquery
- history

In order to handle string requests, we need to understand string when discerning whether
we're dealing with an insert or a retract, so there are a couple updates to handle
that (fql.parse/update? and update/insert? and update/retract?).

In each case we unwrap the credential subject before parsing. We only use a limited
subset of the credential spec internally, but the wide world of credentials has a very
large schema and I didn't want to include it in our internal query/transaction schema.