used derived id for verified credential issuer
We were just pulling the "issuer" field off of the verified credential, but that part of
the credential is not signed. Now we just use the did key from the proof, which is part
of the cryptographic signature.

Also refactored the `generate` function and stubbed out an assertion that we can use in
the future once our incoming credential subjects are proper json-ld.