unwrap db policy after stage
When staging with a policy-activating opt (:did or :role), you need to be able to
subsequently transact with different policies for different users. This commit makes
that possible by resetting the db to a root db after checking policies.

If you want to interact with the db with a specific policy wrapping, you need to specify
the identity/role with which you are doing so every time.