use more efficient query for did roles
Since we know the subject ids for the predicates we care about, we don't need to invoke
the whole query pipeline in order to get the results we care about.

Cleaned up a redundant ns require for fql.

Added some test cases to verify that no errors are thrown when there are no roles or
identities.