Ensure objects of `fluree:targetClass` are correctly handled as types, regardless of txn order
- adds `fluree:targetClass` to list of predefined properties
- extends check for existence of type flakes to make sure that a flake with predicate `rdf:type` and object `rdfs:Class` is created if necessary.

Previously, we just checked if a subject id for a given type already existed, and if so, did not create any flakes for it. But in the case where a type first appears as the object of `fluree:targetClass`, we would create _only_ an "iri flake", so we would have a subject id, but no flake identifying the type as being of type `rdfs:Class`. With this change, we create this additional flake if necessary.