Undo change to create `rdfs:Class` flakes
We have decided to only create these triples when the user explicitly transacts them. This means that if something is _implicitly_ used as a type (eg, `ex:alice` is transacted with `:type` `ex:User`), it may not be returned by queries with clauses like `[?s :rdf/type :rdfs/Class]`, and that is ok.