don't treat sh:class constraint as a sh:nodeKind constraint
This test isn't very high value, but it fails before and passes now.

The problem was treating a "sh:class" property constraint as a "sh:nodeKind"
constraint. When attempting to register the nodekind in our schema cache it would blow
up, because the class sid was not one of the 6 supported nodekinds.