Remove redundant & incorrect test-with-files dep
This shouldn't be in here. We need my fork of it in the cljtest alias for a few things we use sometimes (like keeping the data dir around to inspect the generated files) and if you need it in dev, you should just pull in the cljtest alias too (which you probably want anyway).