only load unindexed commits
There is a race condition when a load that occurs on a ledger whose head commit has
been updated to have an index head behind the data head. Instead of loading just the
unindexed commits, the `trace-commits` function will trace all the way back to the first
commit and attempt to merge them in order. This will fail because the db they are
getting merged into is well past the first transaction.

This commit changes `trace-commits` to only trace back to the first unindexed commit.