add support for sequence paths
This is a simple implementation that uses blocking takes to simplify debugging while
getting it working.

Changes the :path key on a property shape to be a vector of pids if a sequence path is
used. Uses the first pid in the path as the :property key in ths shape map.

If min-count > 0, the last pid in the path is the required pid.

Replaces the given p-flakes with the p-flakes associated with the last pid in the
sequence.

Does not currently handle duplicate properties in the sequence list.