track validated properties in a set
Instead of removing flakes from the data graph, track validated properties in a separate
data structure. This is necessary for cases where the same property may be subject to
multiple constraints (in this case, `property` constraints and `not` constraints).

Also removed some redundant try/catch clauses from the logical tests.