Enable `sh:targetObjectsOf` in cases where nodes are transacted separately, handle datatype
If we have a constraint on `sh:targetObjectsOf` `ex:friend`, we may transact the LHS or the RHS of `ex:friend` in any order, or in separate transactions.

Prior implementation relied on the order [LHS, RHS], and in the same transaction, so that the `subj-mods` atom would contain the necessary shacl data at validation-time.

But if we transacted these subjecsts in separate transactions, the state from the `subj-mods` atom would be blown away in between, and we wouldn't have that data.

With this change, we check to see if a newly-created node might be the RHS of a constrained predicate, so that the `subj-mods` atom can be updated accordingly. We also check in `validate-rules` if there are any subjects we need to validate that are not among those modified directly in the current transaction.

These case also require us to be able to validate datatypes after the triples have already been created. Previously, we only checked datatype constraints at triple-creation time.

For example:

- You have a shape that constraints `sh:targetObjectsOf` `ex:friend`, such that `ex:name` must be a string
- you transact a subject with a non-string `ex:name`. This is valid, if that subject is not yet anyone's friend.
- Then you have a later transaction wherein you do try to make that subject someone’s `ex:friend`

In this case, we cannot rely on the optimization that checks datatypes at triple-creation time, because the `ex:name` triple will have already been created. In cases like these, we need a fall-through case where we can check datatypes during normal shacl-validation operations. This change adds that fall-through case logic.

There are other constraints that similarly need to be enforceable post-triple-creation, and those will be accommodated in a later commit.