add support for sh:node constraints
An sh:node constraint on a property means that the property's object must conform to the
shape specified by sh:node.

In order to validate it, we construct the sh:node shape using the newly abstracted
`build-node-shape` function. Then we resolve the property's object's s-flakes with an
index-range query and then recurse into `validate-shape` with our new shape and our new
s-flakes.