Only look up referring predicates when `sh:targetObjectsOf` rule is active
Because we might add the object of a constrained predicate in a separate transaction from the subject, we need to check which predicates refer to the current node and see if it has any rules on it. If we don't, then we might fail to enforce a pertinent rule. However, we don't want to do this for every node in every txn, since most will not have such a constraint. This way, we do a quick check to see if such a rule is in place before doing the lookup work.