add support for sh:qualifiedValueShapesDisjoint
A qualifiedValueShapesDisjoint constraint in plain English as best as I can:

If you have a shape with a property with a qualifiedValueShape, the set of sibling
shapes is all the other properties' qualifiedValueShapes. When
`sh:qualifiedValueShapesDisjoint` is `true`, then values that conform to that shape
cannot conform to any of the other shapes.

This is tricky becuase it's not a property-shape level concern, we need to compare the
results of multiple property validations. So I broke the whole qualified- family of
constraint checks into its own area.

First, we pass over all property shapes and validate them normally, but now we build up
a collection of qualified-value-shape property shapes (q-shapes) as we go (in
`validate-shape`).

After we've finished checking all of those, we validate the q-shapes as
whole (`validate-q-shapes`).

Here we do the following for each q-shape:
1. build the qualifiedValueShape (qvc)
   note: we only support property shapes here right now
2. add set of conforming sids to each q-shape
3. remove disjoint conforming sids from all other sibling q-shapes
4. validate qualified cardinality constraints