Add fns to generate query schemas with extra k/v pairs
As we're working toward supporting SPARQL-compliant `from` in queries, we are in a state where we are only handling `from` in http-api-gateway, but not adding any logic for it here in db (yet). These fns allow us to add the `from` kw to the schemas for queries in http-api-gateway, without needing to introduce them here yet.