remove multi-query
The primary advantage of multi-query was making it easy to perform multiple queries that are guaranteed to all resolve against the same db commit. We decided to remove it because this can be accomplished in a more straightforward way by getting the `t` first, and then issuing multiple individual queries using that `t`.