add hacky error msg introspection to not break tests
The only non-PropertyShape exception we have is the sh:closed error message, so this
just looks for that message and returns it as-is, and prefixes everything else with the
SHACL PropertyShape exception.

I don't know that this prefix is actually useful, and our validation errors in general
need a complete overhaul, but this preserves the status quo until we decide to do it
properly.