use supplied ledger instead of db ledger
Now we are no longer mutating the db with a commit.

Also changed a few places where we are taking the ledger off a db only to get the conn,
which is also directly on the db.

This is a step towards removing the `:ledger` key from the db entirely.