promote shacl-dt to top-level shape key
Also removed obsolete shacl pipeline.

I needed to add special handling for datatypes because we use the shacl datatype
constraint to coerce values. If the datatype constraint wasn't in a property shape that
was eligible for advanced validation, that coercion wasn't happening. Now we derive the
shacl-dt during shape construction.

Was also able to uncomment old test that depended on post-processing validation.