normalize iri-cache structure and cache subgraph lookups
Before this commit each selector used the iri-cache differently, and therefore gave
broken results if they were combined in a scenario that required iri lookups.

This normalizes the cache structure by making every cache update insert an entry in the
form of `(vswap! iri-cache assoc pid {:as <iri>})`, so that every cache read can use the
results regardless of where it was first inserted.

The other fix in this commit is adding a `dbproto/-iri` lookup to the
`response/wildcard-spec` function. Before it would return a raw pid in the response,
which is unhelpful for users.

fixes #521