add vocabulary caching and loading
This should deprecate the need for pids, though they are still employed. The approach is
to build up the db :schema cache per transaction. When we write out a new index root we
store a vector of all the predicate sids in the root. On load, we query for all the
vocab flakes and recreate the schema cache.

This makes transaction processing a little faster, since we don't have to refresh the
entire vocab cache on every stage.

This also makes loading slower, since we will have to deserialize the (potentially long)
list of predicate sids.