properly merge hydrated schema map into existing one
This is a bit tricky and we should be able to refactor to make it much cleaner in the
future, but this maintains compatibility with the old stage pipeline.

I think it's safe to just discard the old values for :subclasses and :shapes, but I need
additional testing there to be certain.